Upgrade the jrubyStormLocal dependency on storm-core 0.9.2-incubating by default
0.9.2 includes many fixes/improvements and is API incompatible with 0.9.1. The
only requirement that we have for 0.9.1 anymore is on storm topology
deployment, everywhere else we should rely on 0.9.2 if possible

This resolves some frustrating issues with incompatibilities with zookeeper and
curator dependencies underneath storm-kafka 0.9.2-incubating. If you've seen
"cannot initialize class `backtype.storm.LocalCluster` then that means that ZK
dependencies are somehow incorrect in the classpath causing the LocalCluster
class, which is defined in Clojure, to fail to get created. Cute error